Move stringsToRegExp function to app logic
